clean up rxjs use, add material theme
This commit cleans up the Rx calls that were made so you can read them a
bit better. We also more cleanly use the observable onNext function and
set state properly as the users come in through the stream.

This commit also adds Material UI to the project so we can closely mimic
the look of the Android version.

Lastly, we've rearranged the project structure to focus more on
different scenes of the app. A scene folder contains its components and
CSS so far. We also create a folder for common classes, so far only
containing the toolbar component.