introduce container for UserList
This commit pulls the business logic of fetching the users into a
container class. The container makes the call to the injected getUsers
function and then tells the UserList component to render the returned
users.

This means we now have an funciton that can be easily mocked and
separation between business and presentation layers. All that is left is
to write the unit tests!